+++ examples_test.go
@@ -12,7 +12,7 @@ func ExampleClient_newClient() {
 	// Create a new client using an InfluxDB server base URL and an authentication token
 	client := influxdb2.NewClient("http://localhost:9999", "my-token")
 
-	// always close client at the end
+	// Always close client at the end
 	defer client.Close()
 }
 
@@ -22,18 +22,25 @@ func ExampleClient_newClientWithOptions() {
 	client := influxdb2.NewClientWithOptions("http://localhost:9999", "my-token",
 		influxdb2.DefaultOptions().SetBatchSize(20))
 
-	// always close client at the end
+	// Always close client at the end
 	defer client.Close()
 }
 
 func ExampleClient_customServerAPICall() {
+	// Create a new client using an InfluxDB server base URL and empty token
 	client := influxdb2.NewClient("http://localhost:9999", "my-token")
+	// Always close client at the end
+	defer client.Close()
+	// Get generated client for server API calls
 	apiClient := domain.NewClientWithResponses(client.HTTPService())
+	// Get an organization that will own task
 	org, err := client.OrganizationsAPI().FindOrganizationByName(context.Background(), "my-org")
 	if err != nil {
 		//return err
 		panic(err)
 	}
+
+	// Basic task properties
 	taskDescription := "Example task"
 	taskFlux := `option task = {
   name: "My task",
@@ -42,6 +49,8 @@ func ExampleClient_customServerAPICall() {
 
 from(bucket:"my-bucket") |> range(start: -1m) |> last()`
 	taskStatus := domain.TaskStatusTypeActive
+
+	// Create TaskCreateRequest object
 	taskRequest := domain.TaskCreateRequest{
 		Org:         &org.Name,
 		OrgID:       org.Id,
@@ -49,13 +58,19 @@ from(bucket:"my-bucket") |> range(start: -1m) |> last()`
 		Flux:        taskFlux,
 		Status:      &taskStatus,
 	}
+
+	// Issue an API call
 	resp, err := apiClient.PostTasksWithResponse(context.Background(), &domain.PostTasksParams{}, domain.PostTasksJSONRequestBody(taskRequest))
 	if err != nil {
 		panic(err)
 	}
+
+	// Always check generated response errors
 	if resp.JSONDefault != nil {
 		panic(resp.JSONDefault.Message)
 	}
+
+	// Use API call result
 	task := resp.JSON201
 	fmt.Println("Created task: ", task.Name)
 }